M2-D6: JVM bytecode completeness — long/double arrays, multianewarray, wide, goto_w, stack manip
The interpreter's opcode dispatch previously fell through to UnsupportedOpcode
for several opcodes javac legitimately emits — most sharply, any app using
long[] or double[] would crash at runtime. This closes that gap.

Shipped opcodes: laload, daload, lastore, dastore; dup_x1, dup_x2, dup2,
dup2_x1, dup2_x2, swap; multianewarray; wide; goto_w. Deprecated jsr/ret/jsr_w
intentionally skipped (not emitted by javac since Java 6).

Supporting changes:
- ArrayHeap: new load64/store64; long[]/double[] use 2 i32 slots per element;
  length() still reports user-visible count.
- ATYPE_REF arrays gain an ARRAY_TAG bit so aastore/aaload can round-trip
  ArrayRef values (needed for int[][] and friends); GC tracing updated.
- MethodInfo.max_locals / max_stack widened u8 → u16 end-to-end so wide's
  u16 indices are representable.
- pop2 now spec-correct over category-2 tops (single slot, not two).

Tests: 20 new unit tests (6 ArrayHeap 64-bit, 3 long/double array opcodes,
5 stack manipulation incl. cat-2 cases, 2 multianewarray, 4 wide/goto_w).
421/421 passing.

End-to-end: examples/bytecodecoverage/ exercises long[], double[], int[][]
via multianewarray, and StringBuilder.append(long) (which forces dup_x2).

+++ jvm/src/class_file/mod.rs
@@ -51,8 +51,8 @@ pub struct MethodInfo {
     /// 0 means the method is native (no Code attribute).
     pub code_offset: usize,
     pub code_len: usize,
-    pub max_stack: u8,
-    pub max_locals: u8,
+    pub max_stack: u16,
+    pub max_locals: u16,
     pub access_flags: u16,
     /// Exception table parsed from the Code attribute.
     pub exception_table: Vec<ExceptionEntry>,
